+++ R/data.R
@@ -21,7 +21,10 @@
 #'
 #' shows the boundaries of the Local Hospital Network in Australia.
 #' @source <https://hub.arcgis.com/datasets/ACSQHC::local-hospital-networks/explore>
-#' and <https://discover.data.vic.gov.au/dataset/dhhs-service-areas> for Victoria.
+#' for most of Australia but <https://discover.data.vic.gov.au/dataset/dhhs-service-areas>
+#' for Victoria, <https://catalogue.data.wa.gov.au/dataset/health-services-health-006>
+#' for Western Australia, and <https://data.sa.gov.au/data/dataset/local-health-networks>
+#' for South Australia.
 "lhn"