+++ NEWS.md
@@ -1,6 +1,9 @@
 # hpa.spatial.data (development version)
 
-* Initial commit with LHN, PHN and Mesh Blocks data.
+* update LHN shape file to use LHNs for Western Australia from
+  <https://catalogue.data.wa.gov.au/dataset/health-services-health-006>.
 
-* update LHN shape files to use LHNs for South Australia from 
-  <https://data.sa.gov.au/data/dataset/local-health-networks>
+* update LHN shape file to use LHNs for South Australia from 
+  <https://data.sa.gov.au/data/dataset/local-health-networks>.
+
+* Initial commit with LHN, PHN and Mesh Blocks data.